requirements-dev.txt
homeassistant>=2026.3.0
pytest>=8,<10
pytest-cov>=6,<7
python-socketio==5.11.2
ruff>=0.11,<1
mypy>=1.10,<2
mutmut>=3,<4
types-PyYAML>=6,<7
